backend/requirements.txt
anthropic>=0.42.0
py-clob-client>=0.34.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
